src/backend/app/models.py
from newspaper import Article
from newspaper import build
import requests

# Placeholder for fetching news based on interests using Newspaper3k and ChatGPT API
def fetch_news(interests):
    # Example: Fetch news from a specific source, this should be expanded based on your needs
    cnn_paper = build('http://cnn.com')

    # Collect all relevant news articles
    articles = []
    for article in cnn_paper.articles:
        try:
            article.download()
            article.parse()
            articles.append({
                'title': article.title,
                'url': article.url
            })
        except Exception as e:
            print(e)
            continue

    # Placeholder: Send articles and interests to ChatGPT API for filtering
    # filtered_news = call_chatgpt_api(articles, interests)
    filtered_news = articles  # Replace with actual call to ChatGPT

    return filtered_news

# Function to integrate with ChatGPT API (Pseudo-code)
# def call_chatgpt_api(articles, interests):
#     # Implement the logic to call the ChatGPT API and filter articles based on interests
#     return filtered_articles_based_on_interests
