src/backend/requirements.txt
Flask==2.0.1
requests==2.25.1
newspaper3k==0.2.8
python-dotenv==0.19.0
